✨ ユーザ登録とログイン機能の実装

+++ index.js
@@ -3,13 +3,15 @@ const express = require("express")
 const app = express()
 app.use(express.urlencoded({ extended: true }))
 app.use(express.json())
+const jwt = require("jsonwebtoken")
+const auth = require("./utils/auth")
 const connectDB = require("./utils/database")
-const { ItemModel } = require("./utils/schemaModels")
+const { ItemModel, UserModel } = require("./utils/schemaModels")
 
 
 // ITEM functions
 // Create Item
-app.post("/item/create", async(req, res) => {
+app.post("/item/create", auth, async(req, res) => {
     try{
         await connectDB()
         await ItemModel.create(req.body)
@@ -42,22 +44,32 @@ app.get("/item/:id", async(req, res) => {
 })
 
 // Update Item
-app.put("/item/update/:id", async(req, res) => {
+app.put("/item/update/:id", auth, async(req, res) => {
     try{
         await connectDB()
-        await ItemModel.updateOne({_id: req.params.id}, req.body)
-        return res.status(200).json({message: "アイテム編集成功"})
+        const singleItem = await ItemModel.findById(req.params.id)
+        if(singleItem.email === req.body.email){
+            await ItemModel.updateOne({_id: req.params.id}, req.body)
+            return res.status(200).json({message: "アイテム編集成功"})
+        }else{
+            throw new Error()
+        }
     }catch(err){
         return res.status(400).json({message: "アイテム編集失敗"})
     }
 })
 
 // Delete Item
-app.delete("/item/delete/:id", async(req, res) => {
+app.delete("/item/delete/:id", auth, async(req, res) => {
     try{
         await connectDB()
-        await ItemModel.deleteOne({_id: req.params.id})
-        return res.status(200).json({message: "アイテム削除成功"})
+        const singleItem = await ItemModel.findById(req.params.id)
+        if(singleItem.email === req.body.email){
+            await ItemModel.deleteOne({_id: req.params.id})
+            return res.status(200).json({message: "アイテム削除成功"})
+        }else{
+            throw new Error()
+        }
     }catch(err){
         return res.status(400).json({message: "アイテム削除失敗"})
     }
@@ -65,7 +77,40 @@ app.delete("/item/delete/:id", async(req, res) => {
 
 // USER functions
 // Register User
+app.post("/user/register", async(req, res) => {
+    try{
+        await connectDB() 
+        await UserModel.create(req.body)
+        return res.status(200).json({message: "ユーザー登録成功"})
+    }catch(err){
+        return res.status(400).json({message: "ユーザー登録失敗"})
+    }
+})
+
 // Login User
+const secret_key = "tundoku-tower"
+app.post("/user/login", async(req, res) => {
+    try{
+        await connectDB()
+        const savedUserData = await UserModel.findOne({email: req.body.email})
+        if(savedUserData){
+            if(req.body.password === savedUserData.password){
+                const payload = {
+                    email: req.body.email,
+                }
+                const token = jwt.sign(payload, secret_key, {expiresIn: "23h"})
+                console.log(token)
+                return res.status(200).json({message: "ログイン成功", token: token})
+            }else{
+                return res.status(400).json({message: "ログイン失敗：パスワードが間違っています"})
+            }
+        }else{
+            return res.status(400).json({message: "ログイン失敗：ユーザー登録してください"})
+        }
+    }catch(err){
+        return res.status(400).json({message: "ログイン失敗"})
+    }
+})
 
 app.listen(5000, () => {
     console.log("Listening on localhost port 5000")
